Added SpawnBase class and implementation SpawnNormal for standard level (moved code over from StateLevel)
Added scoring of idle movement without kills (not properly ballanced atm -> dependant on spawns, too)

+++ src/ScoreNormal.h
@@ -18,7 +18,7 @@ public:
 
 	void render( SDL_Surface *target );
 protected:
-	int points;
+	float points;
 	int kills;
 	int streak;
 	float multiplier;
